django/app/events/models.py
from django.db import models
import recurrence.fields


class Event(models.Model):
    title = models.CharField(max_length=200)
    start = models.TimeField()
    end = models.TimeField()
    recurrences = recurrence.fields.RecurrenceField()
