django/requirements.txt
Django~=2.1
django-redis-cache~=2.0
django-recurrence~=1.9
psycopg2-binary~=2.7
gunicorn~=19.9
aiohttp~=3.5
aiohttp_wsgi~=0.8
uvloop~=0.12